py/requirements.txt
Cython==0.20.1
Pillow==2.4.0
matplotlib==1.3.1
numpy==1.8.1
pyparsing==2.0.2
pyproj==1.9.3
pyshp==1.2.1
scipy==0.14.0
tornado==3.2.2
wsgiref==0.1.2
